test: rework jest tests
