+++ packages/salesforcedx-vscode-lwc/src/testSupport/types/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, salesforce.com, inc.
+ * Copyright (c) 2023, salesforce.com, inc.
  * All rights reserved.
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
@@ -40,6 +40,15 @@ export const enum TestInfoKind {
   TEST_CASE = 'testCase',
   TEST_FILE = 'testFile',
   TEST_DIRECTORY = 'testDirectory'
+}
+
+/**
+ * Confirms if the TestExecutionInfo is specific to a testCase
+ */
+export function isTestCase(
+  testExecutionInfo: TestExecutionInfo
+): testExecutionInfo is TestCaseInfo {
+  return testExecutionInfo.kind === TestInfoKind.TEST_CASE;
 }
 
 /**
