+++ packages/salesforcedx-vscode-lwc/test/jest/testSupport/testRunner/types.test.ts
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2023, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+import { isTestCase, TestCaseInfo, TestFileInfo, TestInfoKind, TestType } from '../../../../src/testSupport/types';
+import { Uri } from 'vscode';
+import * as path from 'path';
+
+describe('test support types Unit Tests.', () => {
+  const mockUriPath = path.join('/');
+  const mockURI = {
+    fsPath: mockUriPath
+  } as Uri;
+  describe('test isTestCase', () => {
+    it('Should return true for a TestCase', async () => {
+      const testName = 'Testing is Fun!';
+      const testExecutionInfo: TestCaseInfo = {
+        kind: TestInfoKind.TEST_CASE,
+        testType: TestType.LWC,
+        testUri: mockURI,
+        testName
+      };
+      expect(isTestCase(testExecutionInfo)).toBeTruthy();
+    });
+    it('Should return false for a TestFile', async () => {
+      const testExecutionInfo: TestFileInfo = {
+        kind: TestInfoKind.TEST_FILE,
+        testType: TestType.LWC,
+        testUri: mockURI
+      };
+      expect(isTestCase(testExecutionInfo)).toBeFalsy();
+    });
+  });
+});
